Add Pomodoro elapsed time
This replaces the current time to zero counter that only does increments
in a resolution of 1 minute with a second by second updating timer.

You can easily modify the default to count down instead of up (perhaps a
different commit), but I find this to be more granular and helpful for
my Pomodoro sessions.